Use u32 for mode. It has different base size between platforms.

+++ src/main.rs
@@ -192,7 +192,7 @@ fn process_file(abs_path: &str, f: &File) -> Result<FileEntry, String> {
                             owner: an_owner,
                             path: abs_path.to_string(),
                             size: metadata.size(),
-                            mode: metadata.mode(),
+                            mode: metadata.mode() as u32,
                             modified: get_time().sec - metadata.mtime(),
                             .. Default::default()
                         };
